+++ packages/fabric/editor-common/src/schema/nodes/extension.ts
@@ -29,6 +29,7 @@ export const extension: NodeSpec = {
     extensionKey: { default: '' },
     parameters: { default: null },
     text: { default: null },
+    originalContent: { default: null }, // The original Atlassian Document content
   },
   parseDOM: [
     {