+++ packages/fabric/editor-core/src/editor/ui/Appearance/Comment.tsx
@@ -101,6 +101,7 @@ const ContentArea = styled(ContentStyles)`
   overflow-y: auto;
 
   /** Hack for Bitbucket to ensure entire editorView gets drop event; see ED-3294 **/
+  /** Hack for tables controlls. Otherwise marging collapse and controlls are misplaced. **/
   .ProseMirror {
     padding: 12px 20px;
   }
@@ -181,6 +182,7 @@ export default class Editor extends React.Component<
       maxHeight,
       onSave,
       onCancel,
+      disabled,
     } = this.props;
     const maxContentSizeReached =
       maxContentSize && maxContentSize.maxContentSizeReached;
@@ -204,6 +206,7 @@ export default class Editor extends React.Component<
                 items={primaryToolbarComponents}
                 popupsMountPoint={popupsMountPoint}
                 popupsBoundariesElement={popupsBoundariesElement}
+                disabled={disabled}
               />
               <MainToolbarCustomComponentsSlot
                 width={
@@ -224,17 +227,26 @@ export default class Editor extends React.Component<
                 items={contentComponents}
                 popupsMountPoint={popupsMountPoint}
                 popupsBoundariesElement={popupsBoundariesElement}
+                disabled={disabled}
               />
             </ContentArea>
             <SecondaryToolbar>
               <ButtonGroup>
                 {!onSave ? null : (
-                  <Button appearance="primary" onClick={this.handleSave}>
+                  <Button
+                    appearance="primary"
+                    onClick={this.handleSave}
+                    isDisabled={disabled}
+                  >
                     Save
                   </Button>
                 )}
                 {!onCancel ? null : (
-                  <Button appearance="subtle" onClick={this.handleCancel}>
+                  <Button
+                    appearance="subtle"
+                    onClick={this.handleCancel}
+                    isDisabled={disabled}
+                  >
                     Cancel
                   </Button>
                 )}