+++ packages/elements/empty-state/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @atlaskit/empty-state
 
+## 0.1.2
+
 ## 0.1.1
 
 ## 0.1.0