+++ packages/fabric/media-avatar-picker/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @atlaskit/media-avatar-picker
 
+## 4.3.4
+
+## 4.3.3
+
 ## 4.3.2
 
 ## 4.3.1